Use the WP HTTP API methods.
For better backwards and forwards compatibility, use WordPress' HTTP
API. This API will use the best available transport mechanism in the
particular environment. Additionally, if the API is used, maintenance of
the transport is offloaded to WordPress in the event that something
about the transports change.

This change is 1:1 with the previous version using the cURL methods in
PHP. While it may seem that some things were removed (e.g.,
"CURLOPT_RETURNTRANSFER" and "CURLOPT_POSTFIELDS"), these are set by
default in the HTTP API.