Properly enqueue scripts and styles.
WordPress requires that scripts and styles are enqueued on one of the
"enqueue" actions. Not doing so prompts a PHP notice:

`PHP Notice:  wp_register_script was called
<strong>incorrectly</strong>. Scripts and styles should not be
registered or enqueued until the <code>wp_enqueue_scripts</code>,
<code>admin_enqueue_scripts</code>, or
<code>login_enqueue_scripts</code> hooks. Please see <a
href="http://codex.wordpress.org/Debugging_in_WordPress">Debugging in
WordPress</a> for more information. (This message was added in version
3.3.)`

The admin scripts and styles were moved to the correct action to avoid
this notice, as well as prevent possible plugin conflicts.