Address XSS/CSRF vulnerabilities.
Changes admin.php to use the esc_attr() and esc_urL() on the Fastly
option values prior to outputting them in the context of the HTML to be
rendered in the admin panel. While it appears that the values of these
options could not be set from an unprivileged context we should be
escaping them prior to rendering irregardless.

Previously the FastlyAdmin::ajaxSetPage() and FastlyAdmin::ajaxSignUp()
handlers in lib/admin.php were not enforcing the presence of a CSRF
token. This commit also modifies these handlers to use the
wp_verify_nonce() function. The static/fastly.js AJAX posts were
modified to supply a wp_create_nonce() token injected into the scripts
context by wp_localize_script(). The wp_register_script() and associated
calls were moved to a wp_enqueue_scripts event handler in order to
ensure that wp_create_nonce() is defined/loaded at the time we try to
use it.

Thanks to Zack Tollman for the original vulnerability reports.