Further refines security fixes.
Changes the CSRF token to use a localize_script variable that is scoped
with a fastly prefix to avoid collision with any existing variables.

Changes all calls to update_option() with dynamic content to use
esc_sql() on the data before passing it to update_option() to avoid
potential for SQLi.

Escapes content embedded into HTML in the welcome() function to use
esc_attr() and esc_url() to prevent any potential for XSS.