Use trailingslashit for API requests
In both test_fastly_api_connection and the constructor of Vcl_Handler,
the hostname variable was being used but there was nothing done to check
to see if there was a trailing slash or not. In other parts of the
plugin, the function `trailingslashit` was used, so this seemed like a
good way to do it.